src/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stars"
dynamic = ["version"]
description = "智能分析并分类GitHub用户的Starred仓库"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "LileyZhao"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31.0",
    "openai>=1.3.0",
    "rich>=13.4.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0",
    "Jinja2>=3.1.2",
]

[project.urls]
"Homepage" = "https://github.com/username/stars"
"Bug Tracker" = "https://github.com/username/stars/issues"

[tool.setuptools]
packages = ["stars"]
package-dir = {"" = "."}

[tool.setuptools.dynamic]
version = {attr = "stars.__version__"}

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100 